allProducts method will include consumable names as in plist
Actually the removeAllKeychainData method will not delete the remaining count number of the single consumable items.
This happens because remaining consumable items are handled through a custom "name" as expressed in the plist configuration file, while the removeAllKeychainData method just deletes keychain data linked to the product identifier code. So, consumable data will never be completely erased.
Adding a list of the local consumable names to the allProducts solves the problem.